All test and full check of the code

+++ AirWatcher/Controller/Statistics.h
@@ -5,14 +5,14 @@
 class Statistics
 {
 public :
-	double CircularMeanAirQuality(double latitude, double longitude, double radius, time_t* time=nullptr);
+	int CircularMeanAirQuality(double latitude, double longitude, double radius, time_t* time=nullptr);
 
 	double AirQualitySensor(Sensor* sensor, time_t* end = nullptr);
 
 	Statistics(Model* unModel);
 
 private :
-	double atmoIndex(double O3, double So2, double No2, double Pm10);
+	int atmoIndex(double O3, double So2, double No2, double Pm10);
 
 	Model* model;
 };
